Priority Selection made available in SMTP Subject Line
The subject line can be appended with an integer to define the push notification priority (between -2 and 2 per Pushover). Separate the integer and the rest of the subject with a colon.
For example "SUBJECT: My Subject Line:2\n\nMy Email Body" ... this will send as a priority 2 (highest) notification.
High level notifications require the addition of retry and expiration values, the defaults of those values are:
Retry Period: 30 seconds
Expiration Time: 1 hour
Additionally, the default priority is 0 (normal)
These can be changed by setting ENV variables:
PUSHOVER_DEFAULT_RETRY
PUSHOVER_DEFAULT_EXPIRE
PUSHOVER_DEFAULT_PRIORITY

If no priority value is found in the subject line it will use whatever the default priority is.